Loosen test about order of initializers
This commit modifies the code (but not the purpose) of a test that checks that

> initializers are executed after application configuration initializers

Currently the test hard-codes the *exact* initializers that are expected to
occur before a custom one. This can cause the test to fail even if the
expectation still passes.

This commit loosens the test by simply checking that, in the array of
initializers, the custom initializers (called `dummy_initializer` in the
example) is executed after the last occurrence of `load_config_initializers`.